inline tracing logic
