+++ pacesetter-cli/blueprint/web/Cargo.toml.liquid
@@ -19,6 +19,8 @@ serde = { version = "1.0", features = ["derive"] }
 tokio = { version = "1.34", features = ["full"] }
 tower-http = { version = "0.5", features = ["full"] }
 tracing = "0.1"
+tracing-panic = "0.1"
+tracing-subscriber = { version = "0.3", features = ["env-filter", "registry", "fmt"] }
 {% unless template_type == "minimal" -%}
 uuid = { version = "1.6", features = ["serde"] }
 {%- endunless %}
