+++ pacesetter-cli/blueprint/web/src/lib.rs
@@ -5,6 +5,8 @@ use pacesetter::get_env;
 use std::fmt::{Debug, Display};
 use tokio::net::TcpListener;
 use tracing::info;
+use tracing_panic::panic_hook;
+use tracing_subscriber::{filter::EnvFilter, fmt, layer::SubscriberExt, util::SubscriberInitExt};
 
 pub mod controllers;
 pub mod middlewares;
@@ -60,3 +62,22 @@ where
     StatusCode::INTERNAL_SERVER_ERROR
 }
 
+/// Initializes tracing.
+///
+/// This function
+///
+/// * registers a [`tracing_subscriber::fmt::Subscriber`]
+/// * registers a [`tracing_panic::panic_hook`]
+///
+/// The function respects the `RUST_LOG` if set or defaults to filtering spans and events with level [`tracing_subscriber::filter::LevelFilter::INFO`] and higher.
+pub fn init_tracing() {
+    let filter = EnvFilter::try_from_default_env()
+        .or_else(|_| EnvFilter::try_new("info"))
+        .unwrap();
+    tracing_subscriber::registry()
+        .with(fmt::layer())
+        .with(filter)
+        .init();
+
+    std::panic::set_hook(Box::new(panic_hook));
+}
