+++ pacesetter/src/util.rs
@@ -3,8 +3,6 @@ use std::cmp::PartialEq;
 use std::env;
 use std::fmt::{Display, Formatter};
 use tracing::info;
-use tracing_panic::panic_hook;
-use tracing_subscriber::{filter::EnvFilter, fmt, layer::SubscriberExt, util::SubscriberInitExt};
 
 /// The environment the application runs in – either Development, Production, or Test.
 ///
@@ -60,26 +58,6 @@ pub(crate) fn parse_env(env: &str) -> Result<Environment, anyhow::Error> {
         "production" => Ok(Environment::Production),
         unknown => Err(anyhow!(r#"Unknown environment: "{}"!"#, unknown)),
     }
-}
-
-/// Initializes tracing.
-///
-/// This function
-///
-/// * registers a [`tracing_subscriber::fmt::Subscriber`]
-/// * registers a [`tracing_panic::panic_hook`]
-///
-/// The function respects the `RUST_LOG` if set or defaults to filtering spans and events with level [`tracing_subscriber::filter::LevelFilter::INFO`] and higher.
-pub fn init_tracing() {
-    let filter = EnvFilter::try_from_default_env()
-        .or_else(|_| EnvFilter::try_new("info"))
-        .unwrap();
-    tracing_subscriber::registry()
-        .with(fmt::layer())
-        .with(filter)
-        .init();
-
-    std::panic::set_hook(Box::new(panic_hook));
 }
 
 #[cfg(test)]
